refactor(lsp): extract shared offsetOfLine, deduplicate code
- Move offsetOfLine to util.ts and reuse in typeResolution, definition,
  and completion (eliminates 3 redundant source.split calls)
- Use findDefForScope in findLocalDefinition instead of duplicating the
  program.nodes loop
- Clean up onWorkspaceSymbol to use iterator.next() instead of for-return
- Remove redundant WHAT comments
