+++ packages/agency-lang/lib/lsp/typeResolution.ts
@@ -2,6 +2,7 @@ import { getWordAtPosition } from "../cli/definition.js";
 import type { AgencyProgram, VariableType } from "../types.js";
 import type { ScopeInfo } from "../typeChecker/types.js";
 import { findContainingScope } from "./scopeResolution.js";
+import { offsetOfLine } from "./util.js";
 
 export function resolveTypeAtPosition(
   source: string,
@@ -21,14 +22,4 @@ export function resolveTypeAtPosition(
   const resolved = scope.scope.lookup(word);
   if (!resolved || resolved === "any") return null;
   return resolved;
-}
-
-function offsetOfLine(source: string, line: number): number {
-  let offset = 0;
-  for (let i = 0; i < line; i++) {
-    const idx = source.indexOf("\n", offset);
-    if (idx === -1) return source.length;
-    offset = idx + 1;
-  }
-  return offset;
 }
